Making dispatch() public & making _streamListener().
So my reasonings for making dispatch() public & adding the option to 
pass it a prefix is so someone could manually insert MessageEvents into 
an existing CommandsFramework. This would let developers create their 
own streams, for example a guild specific one, and use a prefix for the 
guild that they would pass to the dispatch method. This would prevent 
the need to create a new CommandsFramework for the guild's stream (or 
each guild's stream) with it's specific prefix - which will increase 
memory usage (from my testing). Instead, with this change, one 
CommandsFramework would be made, say listening to a default prefix such 
as a bot mention, and then streams would be made for each guild to 
listen to messages with a specific prefix, and the same logic would 
continue as the one present in _streamListener() when listening to 
messages, but the custom prefix would be passed with dispatch instead of 
no prefix (which results in the CommandsFramework instance prefix to be 
used) - thus allowing for dispatch to work properly.

That somewhat corresponds to branching off _streamListener which was 
branched off to allow listening to continue if onReady has been recieved 
(if a CommandsFramework is made after bot's onReady is recieved - which 
is a pretty niche scenario actually.) It also can present itself as an 
example solution if someone does want to use the public dispatch() 
method - in particular listening to messages in a stream and